refactor(core): consolidate tool-defs/ + executors/ into tools/core/
The built-in core tools were split across two parallel directories —
agent/tool-defs/ (schemas) and agent/executors/ (implementations) —
with every category file paired across both, while every OTHER tool in
agent/tools/ colocates definition + executor in one file. Contributors
had three plausible places to put a new tool.

Each category pair is now ONE file in agent/tools/core/ (defs section +
executors section), with helpers and all tests moved alongside:
- export names unchanged (X_TOOL_DEFS / X_EXECUTORS / CORE_TOOLS /
  CORE_EXECUTORS), so agent/tools.ts re-exports and the agent barrel
  keep their public surface identical
- the two barrel test files become core-defs.test.ts /
  core-executors.test.ts against the combined index
- agent/tool-defs/ and agent/executors/ are deleted
