+++ packages/core/src/agent/tools/core/core-defs.test.ts
