+++ packages/core/src/agent/tools/core/helpers.ts
