+++ packages/core/src/agent/tools/core/resource-tools.test.ts
